requirements.txt
# Daily Puzzles API

# Web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# HTTP client
requests>=2.28.0

# Environment
python-dotenv>=1.0.0

# Google Cloud
google-cloud-firestore>=2.13.0
google-cloud-storage>=2.13.0
google-auth>=2.23.0
